Codes for testing performance
We tested the performance in terms of data reads, especially loading speed and memory footprint. The data set used by our test is too large to be uploaded to github. As an old word says that an uncut gem goes not sparkle. In fact, we prefer to let you use your own data set to test DaPy.